wrap up usable example
