+++ app/openocd.sh
@@ -1,0 +1,1 @@
+openocd -f balena-fin-v1.1.cfg
